Added mount script again

+++ mount.sh
@@ -3,5 +3,5 @@ if ! mountpoint -q /mnt/vdb; then
     mount /dev/vdb /mnt/vdb
 fi
 mkdir -p dataset/
-ln -s /mnt/vdb/Cam2 dataset/ 
-ln -s /mnt/vdb/Cam4 dataset/ 
+[ -L dataset/Cam2 ] || ln -s /mnt/vdb/Cam2 dataset/Cam2
+[ -L dataset/Cam4 ] || ln -s /mnt/vdb/Cam4 dataset/Cam4
